Use a player's email as a authentication hint.
If someone is signed in to Google with multiple accounts at once, they'll have to pick the account to authenticate with every time Google's token expires (which is only after ~1 hour). Giving Google a login hint means players won't have to pick an account after the first time they authenticate (usually when they're using the site for the first time, or if they log out explicitly).